requirements.txt
# FastAPI Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.23

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.1

# AI/ML
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
Pillow>=10.1.0
numpy>=1.26.2

# Utilities
python-dotenv>=1.0.0

# Optional: PostgreSQL support (uncomment if needed)
# psycopg2-binary>=2.9.9

# Testing (optional)
# pytest>=7.4.0
# httpx>=0.24.0
